requirements.txt
alembic==1.4.2
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-HTTPAuth==4.0.0
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.2
MarkupSafe==1.1.1
more-itertools==8.3.0
packaging==20.4
passlib==1.7.2
pkg-resources==0.0.0
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.8.1
pyparsing==2.4.7
pytest==5.4.2
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.17
urllib3==1.25.9
wcwidth==0.1.9
Werkzeug==1.0.1
